Remove the support for the object observer
Move the class DBusObjectObserver to the boss module and rename it
to ModuleObserver. Remove all methods get_observer.
