+++ pyanaconda/dbus/identifier.py
@@ -16,8 +16,6 @@
 # You should have received a copy of the GNU General Public License
 # along with this program.  If not, see <http://www.gnu.org/licenses/>.
 #
-import warnings
-
 from pyanaconda.dbus import DBus
 from pyanaconda.dbus.namespace import get_dbus_path, get_dbus_name
 
@@ -163,20 +161,3 @@ class DBusServiceIdentifier(DBusObjectIdentifier):
         """
         object_path = self._choose_object_path(object_path)
         return self._message_bus.get_proxy(self.service_name, object_path)
-
-    def get_observer(self, object_path=None):
-        """Returns an observer of the DBus object.
-
-        :param object_path: a DBus path of an object or None
-        :return: an observer object
-
-        .. deprecated::
-
-            Use get_proxy instead.
-
-        """
-        warnings.warn("The method get_observer is deprecated. Use get_proxy.",
-                      category=DeprecationWarning, stacklevel=2)
-
-        object_path = self._choose_object_path(object_path)
-        return self._message_bus.get_observer(self.service_name, object_path)
