+++ pyanaconda/modules/boss/module_manager/module_manager.py
@@ -19,6 +19,7 @@
 from pyanaconda.dbus import DBus
 from pyanaconda.dbus.constants import DBUS_START_REPLY_SUCCESS, DBUS_FLAG_NONE
 from pyanaconda.dbus.namespace import get_dbus_name, get_namespace_from_name, get_dbus_path
+from pyanaconda.modules.boss.module_manager import ModuleObserver
 from pyanaconda.modules.common.constants.namespaces import ADDONS_NAMESPACE
 
 from pyanaconda.anaconda_loggers import get_module_logger
@@ -43,7 +44,7 @@ class ModuleManager(object):
         object_path = get_dbus_path(*namespace)
 
         # Add the observer.
-        observer = DBus.get_observer(service_name, object_path)
+        observer = ModuleObserver(DBus, service_name, object_path)
         self._module_observers.append(observer)
 
     def add_addon_modules(self):
